Split Tree into TreeRenderer and TreeNodeRenderer
