+++ packages/hierarchies-react/src/presentation-hierarchies-react.ts
@@ -2,7 +2,9 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-export * from "./presentation-hierarchies-react/Tree";
+
+export { TreeRenderer, getTreeNode } from "./presentation-hierarchies-react/TreeRenderer";
+export { TreeNodeRenderer } from "./presentation-hierarchies-react/TreeNodeRenderer";
 export * from "./presentation-hierarchies-react/Types";
 export * from "./presentation-hierarchies-react/UseTree";
 export { UnifiedSelectionProvider } from "./presentation-hierarchies-react/UnifiedSelectionContext";
